Move some NpmPackageProvider functionality into SimpleWorkerContextProvider
